+++ server/src/test/scala/org/dbpedia/extraction/server/resources/rml/RMLTemplateMappingFactoryTest.scala
@@ -2,8 +2,10 @@ package org.dbpedia.extraction.server.resources.rml
 
 import org.dbpedia.extraction.mappings.MappingsLoader
 import org.scalatest.FunSuite
+
 import scala.language.reflectiveCalls
 import org.dbpedia.extraction.mappings.rml.util.ContextCreator
+import org.dbpedia.extraction.server.resources.rml.model.factories.RMLTemplateMappingFactory
 import org.dbpedia.extraction.util.Language
 import org.dbpedia.extraction.wikiparser.{Namespace, PageNode, WikiParser, WikiTitle}
